Reworked radioLogging protocol and screen into distributedLogging & screen
